Extract ruby version checks into a lib/ library
