+++ lib/entitlements/data/groups/calculated/yaml.rb
@@ -163,7 +163,7 @@ module Entitlements
           # :nocov:
           Contract C::None => C::HashOf[String => C::Any]
           def parsed_data
-            @parsed_data ||= if Entitlements.ruby_version2?
+            @parsed_data ||= if RubyVersionCheck.ruby_version2?
               ::YAML.load(File.read(filename)).to_h
             else
               ::YAML.load(File.read(filename), permitted_classes: [Date]).to_h
