+++ lib/entitlements/data/people/yaml.rb
@@ -79,7 +79,7 @@ module Entitlements
             Entitlements.logger.debug "Loading people from #{filename.inspect}"
 
             # :nocov:
-            raw_person_data = if Entitlements.ruby_version2?
+            raw_person_data = if RubyVersionCheck.ruby_version2?
               ::YAML.load(File.read(filename)).to_h
             else
               ::YAML.load(File.read(filename), permitted_classes: [Date]).to_h
