+++ spec/unit/entitlements/data/groups/calculated/ruby_spec.rb
@@ -103,7 +103,7 @@ describe Entitlements::Data::Groups::Calculated::Ruby do
 
     it "raises an error when an unexpected data structure is created" do
       filename = fixture("ldap-config/filters/filter-bad-data-structure.rb")
-      if Entitlements.ruby_version2?
+      if RubyVersionCheck.ruby_version2?
         expect { described_class.new(filename: filename) }.to raise_error(ParamContractError)
       else
         expect { described_class.new(filename: filename) }.to raise_error(ReturnContractError)
