+++ spec/unit/spec_helper.rb
@@ -1,12 +1,13 @@
 # frozen_string_literal: true
 
+require "ruby-version-check"
 require "base64"
 
 # Note that contracts.ruby has two specific ruby-version specific libraries, which we have vendored into lib/
-if (Gem::Version.new(RUBY_VERSION) >= Gem::Version.new("3.0.0")) # ruby3
-  $LOAD_PATH.unshift(File.expand_path(File.join(__dir__, "../../lib/contracts-ruby3/lib")))
-else # ruby2
+if RubyVersionCheck.ruby_version2?
   $LOAD_PATH.unshift(File.expand_path(File.join(__dir__, "../../lib/contracts-ruby2/lib")))
+else
+  $LOAD_PATH.unshift(File.expand_path(File.join(__dir__, "../../lib/contracts-ruby3/lib")))
 end
 
 require "contracts"
